pyproject.toml
[tool.poetry]
name = "supplierassistant"
version = "0.1.0"
package-mode = false 
description = "LLM-powered assistant that analyzes and answers user queries for supplier-client interactions."
authors = ["Rinat Sharafetdinov <sharafetdinov42@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
langchain-core = "^0.3.49"
langchain-ollama = "^0.3.0"
pydantic = "^2.11.0"
spacy = "^3.8.4"
elasticsearch = "^8.17.2"
pyspellchecker = "^0.8.2"
pandas = "^2.2.3"
langchain-huggingface = "^0.1.2"
scikit-learn = "^1.6.1"
langchain-community = "^0.3.20"
fuzzywuzzy = "^0.18.0"
streamlit = "^1.44.0"
pymongo = "^4.11.3"
langgraph-checkpoint-mongodb = "^0.1.2"
langgraph = "^0.3.21"
python-telegram-bot = "^22.0"
openpyxl = "^3.1.5"
chainlit = "^2.4.400"
openai = "^1.70.0"
llvmlite = "^0.44.0"
numba = "^0.61.0"
openai-whisper = "^20240930"
pymilvus = "^2.5.6"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
ipykernel = "^6.29.5"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 120

[tool.black]
line-length = 120
